requirements.txt
customtkinter
colorama
Pillow
